models/image/npr_deepfakedetection/requirements.txt
flask==3.0.3
pydantic==2.7.0
Pillow==10.3.0
numpy==1.26.4
